rename(slice): rename addBooksSlice to booksSlice
